[fix] プレイヤーの攻撃計算式を修正。武器のパラメータ上昇値が一部間違っていたのを修正。

+++ Assets/Member/Tomiyama/Scripts/WeaponGenerator.cs
@@ -254,8 +254,8 @@ public class WeaponGenerator : MonoBehaviour, IPausable
     {
         public int Damage;
         public int Count;
-        public int AttackSpeed;
-        public int Scale;
+        public float AttackSpeed;
+        public float Scale;
     }
 }
 public enum WeaponType
